Move domain helper logic to redux actions
