+++ packages/colony-starter-react/src/client/actions/domainActions.jsx
@@ -1,21 +1,45 @@
+import BN from 'bn.js'
 import { store } from '../index'
 import * as actions from '../constants/actions'
-import * as domainActions from '../../helpers/actions/domainActions'
 import { getPots } from './fundingActions'
 
 // addDomain
 
 export const addDomain = (colonyClient) => ({
   type: actions.ADD_DOMAIN,
-  payload: domainActions.addDomain(colonyClient)
-    .then(domain => {
-      store.dispatch(getDomains(colonyClient))
-      store.dispatch(getPots(colonyClient))
-      store.dispatch(addDomainSuccess(true))
+  payload: (async () => {
+
+    // add domain
+    const tx = await colonyClient.addDomain.send({
+      parentDomainId: 1,
     })
-    .catch(error => {
-      store.dispatch(addDomainError(error.message))
-    }),
+
+    // check unsuccessful
+    if (!tx.successful) {
+
+      // throw error
+      throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+    }
+
+    // get domain id
+    const { count: domainId } = await colonyClient.getDomainCount.call()
+
+    // get domain
+    const domain = await colonyClient.getDomain.call({ domainId })
+
+    // return domain
+    return domain
+
+  })()
+  .then(domain => {
+    store.dispatch(getDomains(colonyClient))
+    store.dispatch(getPots(colonyClient))
+    store.dispatch(addDomainSuccess(true))
+  })
+  .catch(error => {
+    store.dispatch(addDomainError(error.message))
+  }),
 })
 
 export const addDomainError = (error) => ({
@@ -32,14 +56,41 @@ export const addDomainSuccess = (success) => ({
 
 export const fundDomain = (colonyClient, domainId, amount) => ({
   type: actions.FUND_DOMAIN,
-  payload: domainActions.fundDomain(colonyClient, domainId, amount)
-    .then(success => {
-      store.dispatch(getPots(colonyClient))
-      store.dispatch(fundDomainSuccess(true))
+  payload: (async () => {
+
+    // set token
+    const token = colonyClient.tokenClient.contract.address
+
+    // get domain
+    const { potId } = await colonyClient.getDomain.call({ domainId })
+
+    // move funds between pots
+    const tx = await colonyClient.moveFundsBetweenPots.send({
+      fromPot: 1,
+      toPot: potId,
+      amount: new BN(amount),
+      token,
     })
-    .catch(error => {
-      store.dispatch(fundDomainError(error.message))
-    }),
+
+    // check unsuccessful
+    if (!tx.successful) {
+
+      // throw error
+      throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+    }
+
+    // return true
+    return true
+
+  })()
+  .then(success => {
+    store.dispatch(getPots(colonyClient))
+    store.dispatch(fundDomainSuccess(true))
+  })
+  .catch(error => {
+    store.dispatch(fundDomainError(error.message))
+  }),
 })
 
 export const fundDomainError = (error) => ({
@@ -56,14 +107,45 @@ export const fundDomainSuccess = (success) => ({
 
 export const getDomains = (colonyClient) => ({
   type: actions.GET_DOMAINS,
-  payload: domainActions.getDomains(colonyClient)
-    .then(domains => {
-      store.dispatch(setStateDomains(domains))
-      store.dispatch(getDomainsSuccess(true))
-    })
-    .catch(error => {
-      store.dispatch(getDomainsError(error.message))
-    }),
+  payload: (async () => {
+
+    // get domain count
+    const { count: domainCount} = await colonyClient.getDomainCount.call()
+
+    // set domain id
+    let domainId = 1
+
+    // set domains
+    let domains = []
+
+    // get domains
+    while (domainId <= domainCount) {
+
+      // get domain
+      const domain = await colonyClient.getDomain.call({ domainId })
+
+      // append domain id
+      domain.id = domainId
+
+      // push domain to domains
+      domains.push(domain)
+
+      // increment domain id
+      domainId++
+
+    }
+
+    // return domains
+    return domains
+
+  })()
+  .then(domains => {
+    store.dispatch(setStateDomains(domains))
+    store.dispatch(getDomainsSuccess(true))
+  })
+  .catch(error => {
+    store.dispatch(getDomainsError(error.message))
+  }),
 })
 
 export const getDomainsError = (error) => ({
